Update for so many things

+++ World Server/Buffs.cpp
@@ -98,7 +98,7 @@ bool Buffs::checkBuffs() {
 	return result;
 }
 
-const DWORD Buffs::getStatusAmount(BYTE visualityBit) {
+const WORD Buffs::getStatusAmount(BYTE visualityBit) {
 	for(unsigned int i=0;i<this->buffs.size();i++) {
 		if(this->buffs.getValue(i).visuality == visualityBit)
 			return this->buffs.getValue(i).amount;
